Adding .md generator from .xlsx
We'll need a good way for folks to collaborate on the data that goes
into the media produced here.  Let's use a shared google sheet to hold
things so anyone can contribute.

Here, we add a new generate_markdown.rb utility which can take a
downloaded copy of the sheet made available via a public share link
fetched with Curl in a Make step and populate a new dist/main.md file as
our base for document and website generation.