Adds deployment capability to Makefile
I've got a Digital Ocean droplet running an nginx server to host this
thing now.

Here we add a new deployment target to the Makefile, and add a bit of
documentation to the README since other folks seem interested in
contributing and we want to make things easy for them.